CLOUD-88157 Send notification about resource creation and deletion

+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/service/ReactorFlowManager.java
@@ -59,7 +59,7 @@ import reactor.bus.EventBus;
 @Service
 public class ReactorFlowManager {
 
-    private static final Integer WAIT_FOR_ACCEPT = 5;
+    private static final long WAIT_FOR_ACCEPT = 5L;
 
     @Inject
     private EventBus reactor;
